Fix for ES-934; reorganization
